Use React Redux integration.
It's not necessary to reimplement the wheel in this case. While at it,
wrote the actions with Redux Actions so they're "ok". Also changed
Redux Thunk to Redux Promise instead.

Oh and removed the 'babel-node' startup of the server. Instead opted
to use a bootstrapping file with 'babel/register' in it.